fix issue 267
