+++ sandbox-core/src/main/java/com/alibaba/jvm/sandbox/core/manager/impl/SandboxClassFileTransformer.java
@@ -4,7 +4,6 @@ import com.alibaba.jvm.sandbox.api.event.Event;
 import com.alibaba.jvm.sandbox.api.listener.EventListener;
 import com.alibaba.jvm.sandbox.core.enhance.EventEnhancer;
 import com.alibaba.jvm.sandbox.core.util.ObjectIDs;
-import com.alibaba.jvm.sandbox.core.util.SandboxClassUtils;
 import com.alibaba.jvm.sandbox.core.util.SandboxProtector;
 import com.alibaba.jvm.sandbox.core.util.matcher.Matcher;
 import com.alibaba.jvm.sandbox.core.util.matcher.MatchingResult;
@@ -77,7 +76,7 @@ public class SandboxClassFileTransformer implements ClassFileTransformer {
 
             // 这里过滤掉Sandbox所需要的类|来自SandboxClassLoader所加载的类|来自ModuleJarClassLoader加载的类
             // 防止ClassCircularityError的发生
-            if (SandboxClassUtils.isComeFromSandboxFamily(internalClassName, loader)) {
+            if (SandboxProtector.instance.isComeFromSandboxFamily(internalClassName, loader)) {
                 return null;
             }
